Mostly cleanup
Backup and cache files are no longer versioned.

I've changed the breakout headers to 13 pin from 14 as we don't need
multiple power and ground pins broken out.

I think this board will be much more useful if it has the ability to
talk to a computer, so I'll be adding a part for the FTDI chip in the
reference folder, along with a USB type ? connector.

Currently the board is spaced to fit in a breadboard rather than
compatible with the arduino standard.